+++ server/utils/constants.js
@@ -117,6 +117,10 @@ const EVENTS = {
     CHANGED: 'time.changed',
     SUNRISE: 'time.sunrise',
     SUNSET: 'time.sunset',
+    SUNRISE_PLUS_1: 'time.sunrise-1',
+    SUNRISE_MINUS_1: 'time.sunrise+1',
+    SUNSET_PLUS_1: 'time.sunset-1',
+    SUNSET_MINUS_1: 'time.sunset+1',
   },
   TRIGGERS: {
     CHECK: 'trigger.check',
@@ -466,6 +470,13 @@ const WEBSOCKET_MESSAGE_TYPES = {
     STATUS_CHANGE: 'zigbee2mqtt.status-change',
     MQTT_ERROR: 'zigbee2mqtt.mqtt-error',
     PERMIT_JOIN: 'zigbee2mqtt.permit-join',
+  },
+  ZWAVE2MQTT: {
+    NEW_DEVICE: 'zwave2mqtt.new-device',
+    DISCOVER: 'zwave2mqtt.discover',
+    DISCOVER_COMPLETE: 'zwave2mqtt.discover-complete',
+    STATUS_CHANGE: 'zwave2mqtt.status-change',
+    MQTT_ERROR: 'zwave2mqtt.mqtt-error',
   },
   XIAOMI: {
     NEW_DEVICE: 'xiaomi.new-device',
